backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
supabase>=2.8.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pyjwt>=2.9.0
python-multipart>=0.0.12
bcrypt>=4.2.0
email-validator>=2.2.0
httpx>=0.27.0